:bug: fix NullReferenceException when using both --entitiestokeep and --entitiestoexlude
When using both parameters --entitiestokeep and --entitiestoexlude it can happen that a NullReferenceException is thrown.

This happens when --entitiestokeep does not retain the entities that are then specified to be excluded by --entitiestoexclude. Since the entities were already removed by --entitiestokeep, the logic for --entitiestoexclude then tries to remove entities that were already removed.

#16